Changed resistor values to 12 V
Changed most of the resistors mentioned in these threads:
http://www.electro-music.com/forum/topic-54360.html
http://electro-music.com/forum/topic-36820.html

Not yet changed R23 or R41, in the original schematics. Will do
that after trying it in real life.
Also maybe R46, for fine tuning to span one octave.

Added bill of materials (obviously with the above unchanged values,
that will probably change).