Add panel draft for VCO-1
NOTE: There are hidden layers that can mess up the laser cutting
software (it did for me).

NOTE: The pots in PCB rev A (and probably rev B as well) will have
reversed rotation if mounted as the PCB suggests. In the future I
will mount the pots on the other side (if not fixing the PCB), and
in that case, the panel must be changed (swap the columns of holes).