Use 16 pole power connector
Even though a 10 pole would be enough, my current plan is
to use 16 pole for all boards, thus minimizing cost for
power connectors/cables.

Possibly a bad idea, but if I will go through with the plans
to run workshops using these boards, there will be a LOT of
component bought, and it makes sense to keep cost down.