pro3340: Add BOM, panel, and PCB png
The panel is not tested with PCB, and the BOM is not really
reviewed. E.g. I know it could use better info on the frequency
capacitor (type of capacitor), and the sync switch (it's some odd
kind of three way switch), at least. Also, sockets are not included
(not required, obviously, but might note it so that one doesn't
forget about it).