Moved 3320 VCF to KiCad5 -- NOT VERIFIED
There seem to be a lot of errors with the switches.
Considering replacing them with either jumpers, or possibly
cheaper DPDT push switches (using two for the 3PDT cases, hoping
it won't fry anything).

Still, this commit is just a level up to KiCad 5, no functional
changes done yet. A few will follow.