pro3320: Abandon panel PCB
I tried having a PCB parallell to the panel, but it gets too
wide for my taste (and/or requires too much time and brain
activity in the PCB layout tool).

Will slim it down, having the PCB hanging on the pots, and
mounting the jacks on panel, the switches as well, having
to solder a LOT of cables. It will be hell, but I don't see
another way (except having two PCBs, layered, but that's too
much for me now).