metallic beat: Add jumper for VCA
Mainly for test and experimentation purposes, having a jumper in
the oscillator signal's path to the VCA/filter thing could be nice
to make it possible to run other things through there.

NOTE: When building,one will have to add a capacitor on the jack!
I figured the existing one would be too small to use as a
decoupling capacitor.