Rewrite with classes (with Linux dev support)
Split stuff up in classes, not so much for the code's sake,
but to make it easier to mock the I/O, for development purposes.
So now it's possible to run in Linux, with mouse clicks as taps.

The new main module is lildrum.py, keeping lil_drummer.py around
for a while.